Add files via upload
This project allows the user to change the background color of the entire page by clicking on buttons. When a button is clicked, the background color of the entire page changes to the color associated with that button. The transition is smooth due to the duration-1000 class from Tailwind CSS, which applies a 1-second transition duration.